repair frame size; makes tr work

+++ usr.bin/pmdb/arch/x86_64/x86_64.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD$	*/
+/*	$OpenBSD: x86_64.c,v 1.1 2004/01/29 12:44:40 mickey Exp $	*/
 
 /*
  * Copyright (c) 2002 Federico Schwindt <fgsch@openbsd.org>
@@ -36,8 +36,8 @@
  * No frame for x86?
  */
 struct frame {
-	int fp; 
-	int pc;
+	long fp; 
+	long pc;
 };
 
 static const char *md_reg_names[] = {
